+++ src/components/molecules/AuthForm/useAuthForm.ts
@@ -2,9 +2,9 @@ import { zodResolver } from '@hookform/resolvers/zod'
 import { useMutation } from '@tanstack/react-query'
 import { useContext } from 'react'
 import { SubmitHandler, useForm } from 'react-hook-form'
+import { logInFn, signUpFn } from 'services/auth'
 
 import AuthContext from 'components/shared/AuthContext'
-import { logInFn, signUpFn } from 'config/axios/axios'
 
 import { AuthFormProps, FormInput } from './AuthForm.types'
 import { authSchema } from './authSchema'
@@ -13,7 +13,7 @@ const useAuthForm = ({ isLogInForm = false }: AuthFormProps) => {
   const { logIn } = useContext(AuthContext)
 
   const mutationFn = isLogInForm ? logInFn : signUpFn
-  console.log(mutationFn)
+
   const {
     mutate: performMutation,
     isLoading,
@@ -29,16 +29,13 @@ const useAuthForm = ({ isLogInForm = false }: AuthFormProps) => {
     resolver: zodResolver(authSchema),
   })
 
-  const { reset, handleSubmit } = methods
-
   const onSubmit: SubmitHandler<FormInput> = async (userData) => {
     performMutation(userData)
-    reset()
+    methods.reset()
   }
 
   return {
     methods,
-    handleSubmit,
     onSubmit,
     isLoading,
     isError,
