+++ src/components/molecules/Button/Button.types.ts
@@ -2,11 +2,10 @@ import { ReactNode } from 'react'
 
 export type ButtonProps = {
   children: ReactNode
-  cartButton?: boolean
   primary?: boolean
   secondary?: boolean
-  formButton?: boolean
-  loadingButton?: boolean
+  tertiary?: boolean
+  disabled?: boolean
   type?: 'button' | 'submit'
   onClick?: () => void
 }
