+++ src/components/pages/index.ts
@@ -1,0 +1,8 @@
+export { default as Events } from 'components/pages/Events'
+export { default as Home } from 'components/pages/Home'
+export { default as Layout } from 'components/pages/Layout'
+export { default as LogIn } from 'components/pages/LogIn'
+export { default as Newsletter } from 'components/pages/Newsletter'
+// export { default as ProductDetails } from 'components/pages/ProductDetails'
+export { default as Products } from 'components/pages/Products'
+export { default as SignUp } from 'components/pages/SignUp'
