+++ src/router/index.tsx
@@ -1,13 +1,14 @@
 import { RouteObject } from 'react-router'
 
-import Events from 'components/pages/Events'
-import Home from 'components/pages/Home'
-import Layout from 'components/pages/Layout'
-import Login from 'components/pages/LogIn'
-import Newsletter from 'components/pages/Newsletter'
-// import ProductDetails from 'components/pages/ProductDetails'
-import Products from 'components/pages/Products'
-import SignUp from 'components/pages/SignUp'
+import {
+  Events,
+  Home,
+  Layout,
+  LogIn,
+  Newsletter,
+  Products,
+  SignUp,
+} from 'components/pages'
 
 const routes: RouteObject[] = [
   {
@@ -19,7 +20,7 @@ const routes: RouteObject[] = [
       // { path: 'products/:productName', element: <ProductDetails /> },
       { path: 'events', element: <Events /> },
       { path: 'newsletter', element: <Newsletter /> },
-      { path: 'login', element: <Login /> },
+      { path: 'login', element: <LogIn /> },
       { path: 'signup', element: <SignUp /> },
     ],
   },
